+++ R/calc_bgej.R
@@ -1,112 +1,143 @@
+# calc_bgej
 
-# data-raw/datacreate_bgej.R ?
-# R/calc_bgej.R
-
-################################################################################ #
-
-# > bgej - Create bgej here.####
-
-## old notes
-# blockgroupstats_new_state <- blockgroupstats_new_state[, c("bgid", "bgfips", names_ej, names_ej_supp)]
-# data.table::setDT(blockgroupstats_new_state)
-# data.table::setDT(blockgroupstats_new)
-# data.table::setnames(blockgroupstats_new_state,
-#                      old =  c("bgid", "bgfips", names_ej, names_ej_supp),
-#                      new =  c("bgid", "bgfips", c(names_ej_state, names_ej_supp_state))
-# )
-# # > all.equal(blockgroupstats_new$bgid, blockgroupstats_new_state$bgid)
-# # [1] TRUE
-# data.table::setDF(blockgroupstats_new)
-# data.table::setDF(blockgroupstats_new_state)
-#
-#
-# bgej_new <- data.table::data.table(
-#   blockgroupstats_new[ , c("bgid", "bgfips",
-#                            "ST", "pop",
-#                            names_ej,
-#                            names_ej_supp)],
-#   blockgroupstats_new_state[, c(names_ej_state,
-#                                 names_ej_supp_state)]
-# )
-########################################### # ############################################ # ############################################ #
-
-# use blockgroupstats raw data on Envt indicators and
-# Demog.Index, Demog.
-# and formulas_ejscreen_ej to calculate EJ indexes in every blockgroup
-# as US index, US supplemental (multiple demographics),
-# and State index, State supplemental versions of EJ Index.
-
-
-calc_bgej <- function(dat = blockgroupstats, formulas = formulas_bgej) {
-
-  if (missing(formulas)) {formulas <- formulas_bgej} # $formula and $rname
-  if (is.atomic(formulas) && is.vector(formulas)) {
-    formulas = data.frame(rname = formula_varname(formulas), formula = formulas)
-  } else {
-    if (!all(c("rname", "formula") %in% names(formulas))) {
-      stop("formulas must be a data.frame with elements 'rname' and 'formula' similar to formulas_bgej")
-    }
-  }
-
-  if (missing(dat)) {
-    if (!exists("blockgroupstats")) {
-      stop("must provide dat or else blockgroupstats table must be available")
-    } else {
-      message("Using blockgroupstats dataset for raw envt and demog indicators specified in formulas")
-    }
-    # dat <- blockgroupstats[, c("bgid", "bgfips", "ST", names_e)] # just use formulas to determine what columns are needed
-  }
-
-  if (!("bgfips" %in% names(dat))) {
-    warning("dat should contain at least 'bgfips' column, so this function can calculate State-based EJ indexes and later bgej can be joined to or matched up with blockgroupstats by functions in EJAM that use bgej")
-  }
-  if ("bgfips" %in% names(dat) && !("ST" %in% names(dat))) {
-    dat$ST <- fips2state_abbrev(dat$bgfips)
-  }
-
-  # c(
-  #   "bgid", "bgfips",
-  #   "ST", "pop",
-  #   names_ej,
-  #   names_ej_supp,
-  #   names_ej_state,
-  #   names_ej_supp_state
-  # )
-
-  # > cbind(names_ej, names_ej_supp, names_ej_state, names_ej_supp_state)
-  #      names_ej                          names_ej_supp                       names_ej_state                          names_ej_supp_state
-  # [1,] "EJ.DISPARITY.pm.eo"              "EJ.DISPARITY.pm.supp"              "state.EJ.DISPARITY.pm.eo"              "state.EJ.DISPARITY.pm.supp"
-  # [2,] "EJ.DISPARITY.o3.eo"              "EJ.DISPARITY.o3.supp"              "state.EJ.DISPARITY.o3.eo"              "state.EJ.DISPARITY.o3.supp"
-  # [3,] "EJ.DISPARITY.no2.eo"             "EJ.DISPARITY.no2.supp"             "state.EJ.DISPARITY.no2.eo"             "state.EJ.DISPARITY.no2.supp"
-  # [4,] "EJ.DISPARITY.dpm.eo"             "EJ.DISPARITY.dpm.supp"             "state.EJ.DISPARITY.dpm.eo"             "state.EJ.DISPARITY.dpm.supp"
-  # [5,] "EJ.DISPARITY.rsei.eo"            "EJ.DISPARITY.rsei.supp"            "state.EJ.DISPARITY.rsei.eo"            "state.EJ.DISPARITY.rsei.supp"
-  # [6,] "EJ.DISPARITY.traffic.score.eo"   "EJ.DISPARITY.traffic.score.supp"   "state.EJ.DISPARITY.traffic.score.eo"   "state.EJ.DISPARITY.traffic.score.supp"
-  # [7,] "EJ.DISPARITY.pctpre1960.eo"      "EJ.DISPARITY.pctpre1960.supp"      "state.EJ.DISPARITY.pctpre1960.eo"      "state.EJ.DISPARITY.pctpre1960.supp"
-  # [8,] "EJ.DISPARITY.proximity.npl.eo"   "EJ.DISPARITY.proximity.npl.supp"   "state.EJ.DISPARITY.proximity.npl.eo"   "state.EJ.DISPARITY.proximity.npl.supp"
-  # [9,] "EJ.DISPARITY.proximity.rmp.eo"   "EJ.DISPARITY.proximity.rmp.supp"   "state.EJ.DISPARITY.proximity.rmp.eo"   "state.EJ.DISPARITY.proximity.rmp.supp"
-  # [10,] "EJ.DISPARITY.proximity.tsdf.eo"  "EJ.DISPARITY.proximity.tsdf.supp"  "state.EJ.DISPARITY.proximity.tsdf.eo"  "state.EJ.DISPARITY.proximity.tsdf.supp"
-  # [11,] "EJ.DISPARITY.ust.eo"             "EJ.DISPARITY.ust.supp"             "state.EJ.DISPARITY.ust.eo"             "state.EJ.DISPARITY.ust.supp"
-  # [12,] "EJ.DISPARITY.proximity.npdes.eo" "EJ.DISPARITY.proximity.npdes.supp" "state.EJ.DISPARITY.proximity.npdes.eo" "state.EJ.DISPARITY.proximity.npdes.supp"
-  # [13,] "EJ.DISPARITY.drinking.eo"        "EJ.DISPARITY.drinking.supp"        "state.EJ.DISPARITY.drinking.eo"        "state.EJ.DISPARITY.drinking.supp"
-
-
-dataset_documenter("bgej",
-                   title = "",
-                   description = "bgej is a table of all blockgroups, with the raw scores of the Summary Indexes and Supplemental Summary Indexes for all the environmental indicators.",
-                   details = "This file is not stored in the package.
+#' recalculate EJSCREEN EJ Indexes for all US blockgroups, as for an annual data update
 #'
-#' For documentation on the residential population and environmental data and indicators, see [EJSCREEN documentation](https://web.archive.org/web/20250118193121/https://www.epa.gov/ejscreen/understanding-ejscreen-results).
+#' @param bgstats like [blockgroupstats], a new data.table with one row per blockgroup,
+#'  and columns that include c('bgid', 'bgfips', 'ST', 'pop'),
+#'  environmental indicators with colnames defined in vnames_e,
+#'  and the 4 types of demographic indexes.
+#' @param vnames_e names of columns in bgstats that have envt indicators, assumed if missing
+#' @param vnames_e_pctile optional, just must be same length as vnames_e
+#' @param vnames_e_state_pctile  optional, just must be same length as vnames_e
 #'
-#' The column names are these:
+#' @param vnames_ej optional, names of one set of the EJ Index columns in returned table
+#' @param vnames_ej_supp  optional, names of one set of the EJ Index columns in returned table
+#' @param vnames_ej_state  optional, names of one set of the EJ Index columns in returned table
+#' @param vnames_ej_supp_state  optional, names of one set of the EJ Index columns in returned table
 #'
-#' c('bgfips', 'bgid', 'ST', 'pop',
-#' names_ej,
-#' names_ej_supp,
-#' names_ej_state,
-#' names_ej_supp_state
-#' )" )
+#' @param vnames_d_demogindex name of 1 column in bgstats (that has this 1 type of Demographic Index)
+#' @param vnames_d_demogindex_supp name of 1 column in bgstats
+#' @param vnames_d_demogindex_state name of 1 column in bgstats
+#' @param vnames_d_demogindex_supp_state name of 1 column in bgstats
+#'
+#' @param vnames_ST name of column in bgstats that has the 2-character State abbreviation to use in
+#'   finding envt percentiles in the [statestats] table used by [calc_pctile_columns()]
+#'
+#' @returns data.table like [bgej]
+#'
+#' @export
+#' @keywords internal
+#'
+calc_bgej <- function(bgstats,
+                      vnames_e        = names_e,
+                      vnames_e_pctile       = names_e_pctile,
+                      vnames_e_state_pctile = names_e_state_pctile,
 
+                      vnames_ej       = names_ej,
+                      vnames_ej_supp  = names_ej_supp,
+                      vnames_ej_state      = names_ej_state,
+                      vnames_ej_supp_state = names_ej_supp_state,
 
+                      vnames_d_demogindex      = "Demog.Index",
+                      vnames_d_demogindex_supp = "Demog.Index.Supp",
+                      vnames_d_demogindex_state      = "Demog.Index.State",
+                      vnames_d_demogindex_supp_state = "Demog.Index.Supp.State",
 
-  return(bgej)
+                      vnames_ST = "ST"
+
+) {
+
+  ## GET ENVT DATA AS PERCENTILES
+
+  bge <- data.table::copy(bgstats[, ..vnames_e])
+
+  cat("calculating envt US percentiles\n")
+  epctiles <-
+    calc_pctile_columns(bge,
+                        varnames        = vnames_e,
+                        varnames_pctile = vnames_e_pctile,
+                        zones = "USA")
+
+  cat("calculating envt State percentiles\n")
+  state_epctiles <-
+    calc_pctile_columns(bge,
+                        varnames              = vnames_e,
+                        varnames_state_pctile = vnames_e_state_pctile,
+                        zones = as.vector(unlist(bgstats[, ..vnames_ST])))
+
+  colnames(epctiles)       <- vnames_e_pctile
+  colnames(state_epctiles) <- vnames_e_state_pctile
+
+  data.table::setDT(epctiles)
+  data.table::setDT(state_epctiles)
+
+  ## CALC EJ INDEXES AS ENVT PCTILE X DEMOG INDEX of correct type
+
+  bg <- data.table::copy(bgstats[, c(vnames_d_demogindex, vnames_d_demogindex_supp, vnames_d_demogindex_state, vnames_d_demogindex_supp_state), with = FALSE])
+
+  cat("calculating EJ Indexes from demog indexes and envt percentiles \n")
+
+  bgej_new = cbind(
+    bg[, lapply(.SD, function(z) z * epctiles), .SDcols = vnames_d_demogindex],
+    bg[, lapply(.SD, function(z) z * epctiles), .SDcols = vnames_d_demogindex_supp],
+    bg[, lapply(.SD, function(z) z * state_epctiles), .SDcols = vnames_d_demogindex_state],
+    bg[, lapply(.SD, function(z) z * state_epctiles), .SDcols = vnames_d_demogindex_supp_state]
+  )
+
+  names(bgej_new) <- c(
+    vnames_ej,
+    vnames_ej_supp,
+    vnames_ej_state,
+    vnames_ej_supp_state
+  )
+
+  otheravailable = c('bgid', 'bgfips', 'ST', 'pop')
+  otheravailable = otheravailable[otheravailable %in% names(bgstats)]
+  bgej_new = cbind(bgstats[, ..otheravailable],
+                   bgej_new)
+
+  return(bgej_new)
+}
+################################################################################### #
+################################################################################### #
+if (FALSE) {
+  # ## VALIDATE FOR EXISTING YEAR
+
+  dataload_dynamic("bgej")
+  bgej_old <- bgej[ST %in% c("RI", "DE"), ]
+  bgej_new <- EJAM:::calc_bgej(blockgroupstats[ST %in% c("RI", "DE"), ])
+  #  dim(bgej_old)
+  ## [1] 1498   56
+  #  dim(bgej_new)
+  ## [1] 1498   56
+  # all.equal(bgej_new, bgej_old, check.attributes=FALSE )
+  ## [1] "Column 'EJ.DISPARITY.no2.eo': Mean relative difference: 7.250982e-05"
+  all.equal(bgej_new, bgej_old, check.attributes=FALSE, tolerance = 0.001)
+  ## [1] "Column 'EJ.DISPARITY.drinking.eo': 'is.NA' value mismatch: 285 in current 0 in target"
+
+  EJAM:::nacounts(bgej_old)
+  EJAM:::nacounts(bgej_new)
+  x = data.frame(old = t(bgej_old[bgid == '43899', ]),
+                 new = t(bgej_new[bgid == '43899', ]))
+  x[x$old != x$new & !is.na(x$old), ]
+  ##                            old      new
+  # state.EJ.DISPARITY.pm.eo     0  1.23877
+  # state.EJ.DISPARITY.pm.supp   0 1.510684
+  x[  is.na(x$old), ]
+  ##                                   old new
+  # EJ.DISPARITY.drinking.eo         <NA>   0
+  # EJ.DISPARITY.drinking.supp       <NA>   0
+  # state.EJ.DISPARITY.drinking.eo   <NA>   0
+  # state.EJ.DISPARITY.drinking.supp <NA>   0
+  x[x$new == 0,]
+  ##                                          old new
+  # EJ.DISPARITY.proximity.npdes.eo            0   0
+  # EJ.DISPARITY.drinking.eo                <NA>   0
+  # EJ.DISPARITY.proximity.npdes.supp          0   0
+  # EJ.DISPARITY.drinking.supp              <NA>   0
+  # state.EJ.DISPARITY.proximity.npdes.eo      0   0
+  # state.EJ.DISPARITY.drinking.eo          <NA>   0
+  # state.EJ.DISPARITY.proximity.npdes.supp    0   0
+  # state.EJ.DISPARITY.drinking.supp        <NA>   0
+
 }