+++ R/calc_ejam.R
@@ -1,28 +1,22 @@
 
 
 # calc_ejam()  which uses  calc_byformula()  -  read the formulas and execute them
-# formula_varname()  - utility to get the names of variables created by the formulas
-# formulas_d  - a vector of formulas as text strings like "a = 1 + b" see data-raw/datacreate_formulas.R
-# formulas_ejscreen_acs
+# calc_varname_from_formula()  - utility to get the names of variables created by the formulas
+#   vector of formulas as text strings like "a = 1 + b"
+# formulas_ejscreen_acs$formula
 
 ######################################## #
 
 ## notes
 
-# may want to store formulas for indicators as a table of metadata
-# (rather than formulas being buried in code like doaggregate
-# or even pulled out into a function focused on just that)
-
-# and/or want to allow user to specify a custom indicator or summary stat via formula they provide as text,
-# so EJAM could then aggregate and report the new stat alongside the built-in indicators.
-
-# see datacreate_formulas.R
-
-
+#  store formulas for indicators as a table of metadata but also
+#  want to allow user to specify a custom formula provided as text, to create a new indicator or summary stat
+#  (and EJAM could then aggregate and report the new stat alongside the built-in indicators).
+######################################## #
+## Historical note:
 ## EJAM::calc_ejam() which uses calc_byformula()
 ## was based on
 ## ejscreen::ejscreen.acs.calc() which used analyze.stuff::calc.fields()
-##
 # https://rdrr.io/github/ejanalysis/ejscreen/man/ejscreen.acs.calc.html
 # or more generally here
 #  https://rdrr.io/github/ejanalysis/analyze.stuff/man/calc.fields.html
@@ -31,15 +25,19 @@
 #  https://github.com/ejanalysis/analyze.stuff/blob/36afe6b102cb2cef90b87a48dfea9479b1a2447a/R/calc.fields.R
 #  devtools::install_github("ejanalysis/analyze.stuff")
 #  ?analyze.stuff::calc.fields()
-#
+######################################### #
 # example using just 10 blockgroups from 1 county in Delaware
-#   c1 <- fips2countyname(fips_counties_from_state_abbrev('DE'), includestate = F)[1]
-#   bgdf = data.frame(EJAM::blockgroupstats[ST == "DE" & countyname == c1, ..names_d])[1:10, ]
-#    newdf <-  ejscreen::ejscreen.acs.calc(bgdf, keep.old = "", keep.new = c("my_custom_stat", "mystat2"), formulas = c(
-#      "my_custom_stat <- (pctlowinc + pctmin)/2",
-#      "mystat2  = 100 * pctlowinc"))
-# newdf
-
+if (FALSE) {
+  c1 <- fips2countyname(fips_counties_from_state_abbrev('DE'), includestate = F)[1]
+  bgdf = data.frame(EJAM::blockgroupstats[ST == "DE" & countyname == c1, ..names_d])[1:10, ]
+  newdf <-  ejscreen::ejscreen.acs.calc(
+    bgdf, keep.old = "", keep.new = c("my_custom_stat", "mystat2"),
+    formulas = c(
+      "my_custom_stat <- (pctlowinc + pctmin)/2",
+      "mystat2  = 100 * pctlowinc")
+  )
+  newdf
+}
 ################################################################ #
 
 
@@ -55,7 +53,7 @@
 #'
 #' - [calc_ejam()] uses [calc_byformula()]
 #'
-#' - [calc_byformula()] uses [formula_varname()] and maybe source_this_codetext()
+#' - [calc_byformula()] uses [calc_varname_from_formula()] and maybe source_this_codetext()
 #'
 #' @return data.frame of calculated variables one row per bg row
 #' @examples
@@ -95,7 +93,7 @@
 #' cbind(round(sapply(newdf, max, na.rm=TRUE),2),
 #' names(newdf) %in% names_pct_as_fraction_blockgroupstats)
 #'
-#' EJAM:::formula_varname(formulas_d)
+#' EJAM:::calc_varname_from_formula(formulas_d)
 #'
 #' rm(bgdf)
 #' }
@@ -110,8 +108,8 @@ calc_ejam <- function(bg,
                       quiet = TRUE) {
 
   if (is.null(formulas)) {
-    if (exists("formulas_d")) {
-      formulas <- formulas_d
+    if (exists("formulas_ejscreen_acs")) {
+      formulas <- formulas_ejscreen_acs$formula
     } else {
       warning("no formulas specified or found, so no calculation done")
       return(NULL)
@@ -155,7 +153,7 @@ calc_ejam <- function(bg,
   # don't try to keep fields not supplied in bg
   keep.old <- keep.old[keep.old %in% names(bg)]
   if (keep.new[1] == 'all') {
-    keep.new <- formula_varname(myformulas) # tries to keep all of those formulas would try to create
+    keep.new <- calc_varname_from_formula(myformulas) # tries to keep all of those formulas would try to create
   }
   bg <- calc_byformula(bg, formulas = myformulas, keep = c(keep.old, keep.new), quiet = quiet)
   # if any of these are not successfully created by calc_byformula(), they just won't be returned by that function.
@@ -178,7 +176,7 @@ calc_ejam <- function(bg,
 #' @return data.frame of results, but
 #'   if mydf was a data.table, returns a table in [data.table](https://r-datatable.com) format
 #'
-calc_byformula <- function(mydf, formulas = NULL, keep = formula_varname(formulas), quiet = FALSE) {
+calc_byformula <- function(mydf, formulas = NULL, keep = calc_varname_from_formula(formulas), quiet = FALSE) {
 
 
   # DRAFT WORK NOT COMPLETED
@@ -205,41 +203,41 @@ calc_byformula <- function(mydf, formulas = NULL, keep = formula_varname(formula
   y <- try(source_this_codetext(paste0(formulas, collapse = "; "),  data_list = data_list), silent = TRUE)
   if (inherits(y, "try-error")) {
 
-  for (thisformula in formulas) {
+    for (thisformula in formulas) {
 
-    # *** do we still Need to handle cases where formula tries to use some variable that was not a colname in mydf
-    #   but is in the search path, like in the calling envt. ?
+      # *** do we still Need to handle cases where formula tries to use some variable that was not a colname in mydf
+      #   but is in the search path, like in the calling envt. ?
 
-    # May need to more carefully specify environment, to evaluate in a new envt that contains ONLY the mydf columns? no global envt settings, datasets, etc.??
-  # but some functions might expect and rely on those?
-    # We probably do not want formulas to use variables defined globally or where this function was called from, right?
-    # If not careful about the environment in which things are evaluated,
-    # A problem arises if a formula relies on a variable that is not in mydf but is used by some loaded package,
-    # for example if mydf does not contain the field called cancer but the survival package is attached and provides a dataset lazyloaded called cancer,
-    #  the formula appears to work, but it uses the cancer from the survival package instead of the intended but missing mydf$cancer
-    # So we could check the formula's inputs to see that they are all in mydf,
+      # May need to more carefully specify environment, to evaluate in a new envt that contains ONLY the mydf columns? no global envt settings, datasets, etc.??
+      # but some functions might expect and rely on those?
+      # We probably do not want formulas to use variables defined globally or where this function was called from, right?
+      # If not careful about the environment in which things are evaluated,
+      # A problem arises if a formula relies on a variable that is not in mydf but is used by some loaded package,
+      # for example if mydf does not contain the field called cancer but the survival package is attached and provides a dataset lazyloaded called cancer,
+      #  the formula appears to work, but it uses the cancer from the survival package instead of the intended but missing mydf$cancer
+      # So we could check the formula's inputs to see that they are all in mydf,
 
-    #  or are the outputs of earlier formulas that would work!
+      #  or are the outputs of earlier formulas that would work!
 
-    #    textform <- strsplit(gsub('.*<-', '', parse(text = thisformula)), split=' |[[:punct:]]| ' )[[1]]
-    #    textform <- textform[textform != '']
-    # that gets the words from the right side of the formula, but only if it is written as ....<-....
-    # and it also gets functions like sum, min, max, mean, etc. so it can't really be used to check if each variable is in mydf since some are functions not variables in the formula.
-    # and it splits up on _ etc.
-    ## Also might consider package called  sourcetools that has a tokenize function to parse text into parts
+      #    textform <- strsplit(gsub('.*<-', '', parse(text = thisformula)), split=' |[[:punct:]]| ' )[[1]]
+      #    textform <- textform[textform != '']
+      # that gets the words from the right side of the formula, but only if it is written as ....<-....
+      # and it also gets functions like sum, min, max, mean, etc. so it can't really be used to check if each variable is in mydf since some are functions not variables in the formula.
+      # and it splits up on _ etc.
+      ## Also might consider package called  sourcetools that has a tokenize function to parse text into parts
 
       y <- try(source_this_codetext(thisformula,  data_list = data_list), silent = TRUE)
-if (!quiet) {
-    suppressWarnings(
-      if (inherits(y, "try-error")) {
-        cat('Cannot use formula: '); cat(thisformula, '\n') # print(as.character(parse(text=thisformula)))
-      } else {
-        cat(paste0('Using ', thisformula, '\n'))
+      if (!quiet) {
+        suppressWarnings(
+          if (inherits(y, "try-error")) {
+            cat('Cannot use formula: '); cat(thisformula, '\n') # print(as.character(parse(text=thisformula)))
+          } else {
+            cat(paste0('Using ', thisformula, '\n'))
+          }
+        )
       }
-    )
-      }
+    }
   }
-}
   # RETURN ONLY THE ONES SUCCESSFULLY CREATED, OUT OF ALL REQUESTED TO BE KEPT
 
   keep.from.mydf <- keep[keep %in% names(mydf)]
@@ -272,22 +270,24 @@ if (!quiet) {
 }
 ################################################################ #
 
-
 #' utility to check formulas and extract variable names they calculate values for
 #'
-#' @param myforms see [calc_byformula()]
+#' @param myforms see [calc_byformula()] and [formulas_ejscreen_acs]$formula
 #'
 #' @inherit calc_ejam details
 #'
 #' @examples
-#' EJAM:::formula_varname(c("z=10", "b<- 1", "c <- 34", " h = 1+1", "   q=2+2"))
-#' head(cbind(EJAM:::formula_varname(formulas_d), formulas_d))
+#' EJAM:::calc_varname_from_formula(c("z=10", "b<- 1", "c <- 34", " h = 1+1", "   q=2+2"))
+#' head(cbind(
+#'   EJAM:::calc_varname_from_formula(formulas_ejscreen_acs$formula),
+#'   formulas_ejscreen_acs$formula$formula)
+#'   )
 #'
 #' @return a vector as long as myforms input vector
 #'
 #' @keywords internal
 #'
-formula_varname <- function(myforms) {
+calc_varname_from_formula <- function(myforms) {
 
   return(
     gsub("^([^ <=]*)[ <=].*", "\\1", trimws(myforms))