+++ R/utils_varinfo.R
@@ -27,7 +27,7 @@
 #' - EJAM names_whichlist()
 #' - EJAM names_whichlist_multi()
 #' - EJAM names_whichlist_multi_key()
-#' - EJAM formula_varname()
+#' - EJAM calc_varname_from_formula()
 #'
 #' and the various related data objects like map_headernames and namez
 #'