+++ NEWS.md
@@ -4,7 +4,7 @@
 - v2.4.0, this release, is just to provide 2019-2023 ACS for the EJSCREEN demographics indicators (and lead paint indicator), in case those are useful.
 - v2.5.0 will provide 2020-2024 ACS and will be released in early 2026.
 - Note that Census Bureau discourages using ACS 5-year surveys for comparisons or trends if they have overlapping periods. Comparisons between the 2018-2022, 2019-2023, and 2020-2024 datasets is not appropriate per Census Bureau.
-
+- Functions created to update ACS-based demographic dataset each year: calc_bgej(), calc_blockgroupstats_acs(), calc_blockgroupstats_from_tract_data(), calc_bgwts_nationwide(), etc.
 
 # EJAM 2.32.6.003 (November 2025)